fix(ingestion): staff parser-architecture review — H1/H2/H3/H4/H6/H7/H8/H9, R6 (2.419.0)
H2 script rows keep the scan that first recorded them (a failed rescan's
cleanup cascaded into earlier evidence); H7 equal-confidence rescans refresh
product/version; H3 web parsers isolate each record (record_savepoint) and
sizes are BIGINT; H1 dedup caches drop rolled-back rows; H4 undecodable JSON
lines are counted and truncated masscan XML is partial; H6 a cancelled job
names the scan it wrote; H8 one is_active rule; H9 IPv6 in address fields;
R6 dead NetExec patterns removed.

+++ backend/app/db/models.py
@@ -597,7 +597,9 @@ class WebInterface(Base):
     status_code = Column(Integer)
     title = Column(String)
     server_header = Column(String)
-    content_length = Column(Integer)
+    # BigInteger (v2.419.0, review H3): a body past 2 GiB is real output, and
+    # the 32-bit column failed its flush.
+    content_length = Column(BigInteger)
     # Flattened technology list for UI chip rendering (e.g.
     # ["Nginx 1.18.0", "React", "Bootstrap"]).  Wappalyzer categories
     # preserved in raw.  Null when the tool didn't report any.
@@ -676,7 +678,7 @@ class WebPath(Base):
     url = Column(String, nullable=False)                 # the full URL requested
     path = Column(String, nullable=False, index=True)    # its path, for search / filtering
     status_code = Column(Integer, index=True)
-    size = Column(Integer)
+    size = Column(BigInteger)   # v2.419.0 — as web_interfaces.content_length
     first_seen = Column(DateTime(timezone=True), server_default=func.now())
 
     __table_args__ = (
